0.42

+++ dataset.py
@@ -117,18 +117,18 @@ def collate_fn(insts, opt):
 
     src_pos_list = []
     for src in src_seqs_list:
-        src_pos_list.append(np.array([
-            [pos_i+1 if w_i != constant.PAD else 0
-                for pos_i, w_i in enumerate(inst)]
-                    for inst in src]))
-
+        src_pos_list.append(
+            torch.LongTensor(np.array([
+                    [pos_i+1 if w_i != constant.PAD else 0
+                        for pos_i, w_i in enumerate(inst)]
+                            for inst in src])))
     tgt_pos_list = []
     for tgt in tgt_seqs_list:
-        tgt_pos_list.append(np.array([
-            [pos_i+1 if not(np.array_equal(w_i, np.array([constant.PAD] * w_i.shape[0]))) else 0
-                for pos_i, w_i in enumerate(inst)]
-                    for inst in tgt]))
+            tgt_pos_list.append(
+                torch.LongTensor(np.array([
+                    [pos_i+1 if not(np.array_equal(w_i, np.array([constant.PAD] * w_i.shape[0]))) 
+                            else 0
+                                for pos_i, w_i in enumerate(inst)]
+                                    for inst in tgt])))
                 
-    
-
     return zip(src_seqs_list, src_lens_list, src_pos_list, tgt_seqs_list, tgt_pos_list)
